GH-25 add basic request logging to the server
this is helpful when diagnosing problems with the client-server
connection to see what requests the server is receiving and
what status it is returning.

this also can serve as an example for anyone using this
package how they might gain some visibility into the client-server
communication from the server side.

in a more complex application the LogRequests middleware
might use a logger rather than writing directly to stdout
but for the purpose of this sample application stdout is
sufficient