GH-25 embed the ./public folder
go1.16 introduced the go:embed directive which allows easy
embedding of static files into the built binary.

this is helpful as it simplifies running the sample

the easiest way to run this sample app is

    go run chatsample/*.go

but behind the scenes go is compiling that package
into a temp folder and running the compiled binary
from there.  unfortunately, it was doing so without
any knowledge of the chatsample/public/* content
so I was constantly getting a 404 Not Found when
I loaded up http://localhost:8086

by embedding the public content into the an embed.FS
and serving our static content from there, we are able
to see it through a go run or when running the output
from a go build