server.MapHTTP through an interface
this abstraction removes the explicit binding to http.ServeMux
and allows use of any mux which implements those basic Handle
and HandleFunc methods.

any mux implementation which implements those building blocks
can be used as a MappableRouter and passed into MapHTTP

I tested this with the chatsample and it is backwards compatable
in the sense that go is able to see http.ServeMux as a
MappableRouter without any code changes.