avoid writing superflous response headers
I have noticed that in the server-side logs I sometimes see this:

    2021/07/11 18:26:26 http: superfluous response.WriteHeader call from
      github.com/philippseith/signalr.(*httpMux).handleWebsocket (httpmux.go:118)

On investigation of the httpMux.handleWebsocket method it appears
that the websocket.Accept method writes an error header on every
path which returns an error object; thus when by the time this
method tests for an error returned from websocket.Accept, the
header has already been written.

this commit replaces that superflous writeheader call with a
log statement.