fix: WebSocketConnection: Timeout on Read shuts down writing connection
webSocketConnection Read and Write had separate
timeouts which, when expired, mutually closed the
websocket. In case only one party of the connection
was sending, the websocket was closed after the
timeout despite the fact that the connection was alive.
This comes from misunderstanding the impact of the Read/Write context parameter. See https://github.com/nhooyr/websocket/issues/242